also verify the authName when verifying authTokens

+++ lib/etc/verifyAuthToken.php
@@ -1,11 +1,11 @@
 <?php
 
-    function verifyAuthToken($tokenToVerify) {
+    function verifyAuthToken(string $tokenToVerify, string $authName) {
         require_once dirname(__FILE__)."/../table/authToken.php";
         $token = new authToken($_POST['authToken']);
 
         require_once dirname(__FILE__)."/getClientIpAddress.php";
-        if ($token->existed && $token->dateTimeUsed === NULL && $token->clientIp == getClientIpAddress()) {
+        if ($token->existed && $token->dateTimeUsed === NULL && $token->authName == $authName && $token->clientIp == getClientIpAddress()) {
             if ($token->delete()) {
                 return true;
             }
